requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-Login==0.5.0
Pillow==8.2.0